Created the application for Exam registration
