+++ edu_hub/settings.py
@@ -44,6 +44,7 @@ INSTALLED_APPS = [
     'rest_framework_simplejwt',
     'users',
     'academics',
+    'ExamRegistration',
     'career',
     'drf_yasg',
 ]
