+++ edu_hub/urls.py
@@ -39,6 +39,7 @@ urlpatterns = [
     path('api/auth/', include('users.urls')),
     path('api/academics/', include('academics.urls')),
     path('api/career/', include('career.urls')),
+    path('api/exams/', include('ExamRegistration.urls')),
     re_path(r'^swagger(?P<format>\\.json|\\.yaml)$', schema_view.without_ui(cache_timeout=0), name='schema-json'),
     path('swagger/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
     path('redoc/', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
